collapse helper funcs into api file
